Add quantity component to cart modal
