+++ src/components/Cart/Cart.jsx
@@ -2,9 +2,10 @@ import React from "react";
 import { useState, useEffect } from "react";
 import "./Cart.scss";
 import CartItem from "../cartItem/cartItem";
+// import QuantityPicker from "../QuantityPicker/QuantityPicker";
 
 const Cart = () => {
-  // const [count, setCount] = useState(0);
+//   const [count, setCount] = useState(0);
   const [cartContents, setCartContents] = useState(() => {
     const savedCart = sessionStorage.getItem("cartContents");
     return savedCart ? JSON.parse(savedCart) : [];
@@ -37,30 +38,12 @@ const Cart = () => {
     window.location.reload();
   };
 
-    const increment = (e) => {
-      e.preventDefault();
-      setCount(function (prevCount) {
-        return (prevCount += 1);
-      });
-    };
-
-    const decrement = (e) => {
-      e.preventDefault();
-      setCount(function (prevCount) {
-        if (prevCount > 0) {
-          return (prevCount -= 1);
-        } else {
-          return (prevCount = 0);
-        }
-      });
-    };
-
-    // const handleClick = (e) => {
-    //   e.preventDefault();
-    //   if (count > 0) {
-    //     addToCart({ productId: id, quantity: count, price: price });
-    //   }
-    // };
+  // const handleClick = (e) => {
+  //   e.preventDefault();
+  //   if (count > 0) {
+  //     addToCart({ productId: id, quantity: count, price: price });
+  //   }
+  // };
 
   return (
     <section className="cart">
@@ -93,10 +76,10 @@ const Cart = () => {
       </ul>
       <div className="cart__footer">
         {cartContents.length > 0 ? (
-            <div className="cart__total">
-                <h5 className="cart__total-text">{`Total: `}</h5>
-                <h5 className="cart__total-cost">{`$${calcTotal().toLocaleString()}`}</h5>
-            </div>
+          <div className="cart__total">
+            <h5 className="cart__total-text">{`Total: `}</h5>
+            <h5 className="cart__total-cost">{`$${calcTotal().toLocaleString()}`}</h5>
+          </div>
         ) : (
           <h5 className="cart__total cart__total--empty">{""}</h5>
         )}
